Fixing failing test.
